+++ mcr/src/readings/positions.hpp
@@ -41,15 +41,15 @@ typedef class positionsReading positionsReading_t;
 
 class positionsReading : public Reading {
 private:
-  static const uint16_t _max_pios = 16;
+  static const uint8_t _max_pios = 16;
   // actual reading data
   uint8_t _pios = 0;
-  uint16_t _states = 0x0000;
+  uint8_t _states = 0x0000;
 
 public:
   // undefined reading
-  positionsReading(mcrDevID_t &id, time_t mtime, uint16_t states, uint8_t pios);
-  uint16_t state() { return _states; }
+  positionsReading(mcrDevID_t &id, time_t mtime, uint8_t states, uint8_t pios);
+  uint8_t state() { return _states; }
 
 protected:
   virtual void populateJSON(JsonObject &root);
